requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
langchain>=0.0.340
langchain-openai>=0.0.2
langchain-community>=0.0.11
langgraph>=0.0.15
pydantic>=2.5.2
python-dotenv>=1.0.0